+++ tests/unit/Transformation/ArraySelectionTransformerTest.php
@@ -1,5 +1,6 @@
 <?php
 
+use paslandau\DataFiltering\Exceptions\TransformationException;
 use paslandau\DataFiltering\Transformation\ArraySelectionTransformer;
 
 class ArraySelectionTransformerTest extends PHPUnit_Framework_TestCase
@@ -60,7 +61,7 @@ class ArraySelectionTransformerTest extends PHPUnit_Framework_TestCase
 
     public function testIndexOutOfRange()
     {
-        $this->setExpectedException(get_class(new UnexpectedValueException ()));
+        $this->setExpectedException(TransformationException::class);
         $tests = array(
             [["foo", "bar", "baz", ""], [1, 2, 10]],
         );